[CryptoTracker] - Create Networking service to get coins

+++ SwiftUICrypto/SwiftUICrypto/Models/CoinModel.swift
@@ -59,7 +59,7 @@ struct CoinModel: Identifiable, Codable {
     let priceChange24H, priceChangePercentage24H: Double?
     let marketCapChange24H: Double?
     let marketCapChangePercentage24H: Double?
-    let circulatingSupply, totalSupply, maxSupply, ath: Int?
+    let circulatingSupply, totalSupply, maxSupply, ath: Double?
     let athChangePercentage: Double?
     let athDate: String?
     let atl, atlChangePercentage: Double?
